//VJ 170303 Added concentrate diffusive wave direction, fixed bug in water height output to screen

+++ include/model.h
@@ -789,7 +789,7 @@ public:
     double K2DQSOut;
     double K2DQPOut;
     double CourantKin;
-    int ConcentrateKin;
+    double ConcentrateKin;
     double TimestepKinMin;
     double KinematicBoundaryFraction = 0.05;
     //SWATRE
